+++ src/pages/main/index.tsx
@@ -2,7 +2,6 @@ import { useAppDispatch } from '../../hooks/reduxHooks';
 import { setIsAuth } from '../../redux/userSlice';
 import * as Icons from '../../img/iconsComponents';
 import AddBtn from '../../components/addBtn';
-import OnBoarding from '../onBoarding/onBoarding';
 
 export default function Main() {
   const dispatch = useAppDispatch();
@@ -51,7 +50,6 @@ export default function Main() {
         <Icons.MenuSettings active />
       </div>
       <AddBtn />
-      <OnBoarding />
     </div>
   );
 }
