--- src/pages/onBoarding/styles.css
@@ -1,28 +1,0 @@
-.onboarding__container {
-  @apply flex mx-auto  w-[22.5rem] h-[45rem] bg-[url('../../img/images/1.png')] bg-no-repeat;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.onboarding__wrapper {
-  @apply flex;
-  justify-content: space-between;
-}
-
-.onboarding__button_wrapper {
-  @apply flex gap-1;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.onboarding__content {
-  @apply px-4 w-[100%] pt-16 pb-[72px] bg-[url('../../img/images/onboarding_1.png')] bottom-0;
-}
-
-.onboarding__content_wrapper {
-  @apply mt-4 text-xl text-[#5B5B69];
-}
-
-.onboarding__content_title {
-  @apply font-medium text-4xl text-[#040C0C];
-}
